+++ src/main/java/com/springboot/moa/user/model/GetUserInterestRes.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @AllArgsConstructor
 
 public class GetUserInterestRes {
-    int postId;
+    long postId;
     int point;
     String title;
     // 등록된 질문의 개수
